+++ jitterentropy.h
@@ -42,6 +42,12 @@
 #ifndef _JITTERENTROPY_H
 #define _JITTERENTROPY_H
 
+#if defined(_MSC_VER)
+#include "arch/jitterentropy-base-windows.h"
+#else
+#include "jitterentropy-base-user.h"
+#endif
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -70,138 +76,92 @@ extern "C" {
 		      JENT_MINVERSION * 10000 + \
 		      JENT_PATCHLEVEL * 100)
 
-/***************************************************************************
- * Jitter RNG Configuration Section
- *
- * You may alter the following options
- ***************************************************************************/
+/* -- BEGIN Main interface functions -- */
+/* Flags that can be used to initialize the RNG */
+#define JENT_DISABLE_STIR (1<<0) 	/* UNUSED */
+#define JENT_DISABLE_UNBIAS (1<<1) 	/* UNUSED */
+#define JENT_DISABLE_MEMORY_ACCESS (1<<2) /* Disable memory access for more
+					     entropy, saves MEMORY_SIZE RAM for
+					     entropy collector */
+#define JENT_FORCE_INTERNAL_TIMER (1<<3)  /* Force the use of the internal
+					     timer */
+#define JENT_DISABLE_INTERNAL_TIMER (1<<4)  /* Disable the potential use of
+					       the internal timer. */
+#define JENT_FORCE_FIPS (1<<5)		  /* Force FIPS compliant mode
+					     including full SP800-90B
+					     compliance. */
+#define JENT_NTG1 (1<<6) /* AIS 20/31 NTG.1 compliance */
 
-/*
- * Enable timer-less timer support with JENT_CONF_ENABLE_INTERNAL_TIMER
- *
- * In case the hardware is identified to not provide a high-resolution time
- * stamp, this option enables a built-in high-resolution time stamp mechanism.
- *
- * The timer-less noise source is based on threads. This noise source requires
- * the linking with the POSIX threads library. I.e. the executing environment
- * must offer POSIX threads. If this option is disabled, no linking
- * with the POSIX threads library is needed.
- */
+/* Flags field limiting the amount of memory to be used for memory access */
+#define JENT_FLAGS_TO_MEMSIZE_SHIFT	28
+#define JENT_FLAGS_TO_MAX_MEMSIZE(val)	(val >> JENT_FLAGS_TO_MEMSIZE_SHIFT)
+#define JENT_MAX_MEMSIZE_TO_FLAGS(val)	(val << JENT_FLAGS_TO_MEMSIZE_SHIFT)
+#define JENT_MAX_MEMSIZE_32kB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 1))
+#define JENT_MAX_MEMSIZE_64kB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 2))
+#define JENT_MAX_MEMSIZE_128kB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 3))
+#define JENT_MAX_MEMSIZE_256kB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 4))
+#define JENT_MAX_MEMSIZE_512kB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 5))
+#define JENT_MAX_MEMSIZE_1MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 6))
+#define JENT_MAX_MEMSIZE_2MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 7))
+#define JENT_MAX_MEMSIZE_4MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 8))
+#define JENT_MAX_MEMSIZE_8MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 9))
+#define JENT_MAX_MEMSIZE_16MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C(10))
+#define JENT_MAX_MEMSIZE_32MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C(11))
+#define JENT_MAX_MEMSIZE_64MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C(12))
+#define JENT_MAX_MEMSIZE_128MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C(13))
+#define JENT_MAX_MEMSIZE_256MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C(14))
+#define JENT_MAX_MEMSIZE_512MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C(15))
+#define JENT_MAX_MEMSIZE_MAX		JENT_MAX_MEMSIZE_512MB
+#define JENT_MAX_MEMSIZE_MASK		JENT_MAX_MEMSIZE_MAX
+/* We start at 32kB -> offset is log2(32768) */
+#define JENT_MAX_MEMSIZE_OFFSET		14
 
-/*
- * Shall the LAG predictor health test be enabled?
- */
-#define JENT_HEALTH_LAG_PREDICTOR
-
-/*
- * Shall the jent_memaccess use a (statistically) random selection for the
- * memory to update?
- */
-#define JENT_RANDOM_MEMACCESS
-
-/*
- * Mask specifying the number of bits of the raw entropy data of the time delta
- * value used for the APT.
- *
- * This value implies that for the APT, only the bits specified by
- * JENT_APT_MASK are taken. This was suggested in a draft IG D.K resolution 22
- * provided by NIST, but further analysis
- * (https://www.untruth.org/~josh/sp80090b/CMUF%20EWG%20Draft%20IG%20D.K%20Comments%20D10.pdf)
- * suggests that this truncation / translation generally results in a health
- * test with both a higher false positive rate (because multiple raw symbols
- * map to the same symbol within the health test) and a lower statistical power
- * when the APT cutoff is selected based on the apparent truncated entropy
- * (i.e., truncation generally makes the test worse). NIST has since withdrawn
- * this draft and stated that they will not propose truncation prior to
- * health testing.
- * Because the general tendency of such truncation to make the health test
- * worse the default value is set such that no data is masked out and this
- * should only be changed if a hardware-specific analysis suggests that some
- * other mask setting is beneficial.
- * The mask is applied to a time stamp where the GCD is already divided out, and thus no
- * "non-moving" low-order bits are present.
- */
-#define JENT_APT_MASK		(UINT64_C(0xffffffffffffffff))
-
-/* This parameter establishes the multiplicative factor that the desired
- * memory region size should be larger than the observed cache size; the
- * multiplicative factor is 2^JENT_CACHE_SHIFT_BITS.
- * Set this to 0 if the desired memory region should be at least as large as
- * the cache. If one wants most of the memory updates to result in a memory
- * update, then this value should be at least 1.
- * If the memory updates should dominantly result in a memory update, then
- * the value should be set to at least 3.
- * The actual size of the memory region is never larger than requested by
- * the passed in JENT_MAX_MEMSIZE_* flag (if provided) or JENT_MEMORY_SIZE
- * (if no JENT_MAX_MEMSIZE_* flag is provided).
- */
-#ifndef JENT_CACHE_SHIFT_BITS
-#define JENT_CACHE_SHIFT_BITS 0
-#endif
-
-/*
- * Meory access loop count: This value defines the default memory access loop
- * count. The memory access loop is one of the hearts of the Jitter RNG. The
- * number of loop counts has a direct impact on the entropy rate.
- *
- * It is permissible to configure this value differently at compile time if the
- * observed entropy rate is too small.
- *
- * NOTE: When you modify this value, you are directly altering the behavior of
- * the noise source. Make sure you fully understand what you do. If you want to
- * individually measure the memory access loop entropy rate, use the
- * jitterentropy-hashtime tool with the command line option of --memaccess.
- */
-#ifndef JENT_MEM_ACC_LOOP_DEFAULT
-#define JENT_MEM_ACC_LOOP_DEFAULT 1
-#endif
-
-/*
- * Hash loop count: This value defines the default hash loop count. The hash
- * loop is one of the hearts of the Jitter RNG. The number of loop counts has a
- * direct impact on the entropy rate.
- *
- * It is permissible to configure this value differently at compile time if the
- * observed entropy rate is too small.
- *
- * NOTE: When you modify this value, you are directly altering the behavior of
- * the noise source. Make sure you fully understand what you do. If you want to
- * individually measure the memory access loop entropy rate, use the
- * jitterentropy-hashtime tool with the command line option of --hashloop.
- */
-#ifndef JENT_HASH_LOOP_DEFAULT
-#define JENT_HASH_LOOP_DEFAULT 1
-#endif
-
-/***************************************************************************
- * Jitter RNG State Definition Section
- ***************************************************************************/
-
+#ifdef JENT_PRIVATE_COMPILE
+# define JENT_PRIVATE_STATIC static
+#else /* JENT_PRIVATE_COMPILE */
 #if defined(_MSC_VER)
-#include "arch/jitterentropy-base-windows.h"
+#define JENT_PRIVATE_STATIC __declspec(dllexport)
 #else
-#include "jitterentropy-base-user.h"
+#define JENT_PRIVATE_STATIC __attribute__((visibility("default")))
+#endif
 #endif
 
-#define JENT_SHA3_256_SIZE_DIGEST_BITS	256
-#define JENT_SHA3_256_SIZE_DIGEST	(JENT_SHA3_256_SIZE_DIGEST_BITS >> 3)
-#define JENT_SHA3_512_SIZE_DIGEST_BITS	512
-#define JENT_SHA3_512_SIZE_DIGEST	(JENT_SHA3_512_SIZE_DIGEST_BITS >> 3)
+/* Forward declaration of opaque value */
+struct rand_data;
+
+/* Number of low bits of the time value that we want to consider */
+/* get raw entropy */
+JENT_PRIVATE_STATIC
+ssize_t jent_read_entropy(struct rand_data *ec, char *data, size_t len);
+JENT_PRIVATE_STATIC
+ssize_t jent_read_entropy_safe(struct rand_data **ec, char *data, size_t len);
+/* initialize an instance of the entropy collector */
+JENT_PRIVATE_STATIC
+struct rand_data *jent_entropy_collector_alloc(unsigned int osr,
+	       				       unsigned int flags);
+/* clearing of entropy collector */
+JENT_PRIVATE_STATIC
+void jent_entropy_collector_free(struct rand_data *entropy_collector);
+
+/* initialization of entropy collector */
+JENT_PRIVATE_STATIC
+int jent_entropy_init(void);
+JENT_PRIVATE_STATIC
+int jent_entropy_init_ex(unsigned int osr, unsigned int flags);
 
 /*
- * The output 256 bits can receive more than 256 bits of min entropy,
- * of course, but the 256-bit output of SHA3-512_256(M) / XDRBG-256(M) can only
- * asymptotically approach 256 bits of min entropy, not attain that bound.
- * Random maps will tend to have output collisions, which reduces the creditable
- * output entropy (that is what SP 800-90B Section 3.1.5.1.2 attempts to bound).
- *
- * The value "64" is justified in Appendix A.4 of the current 90C draft,
- * and aligns with NIST's in "epsilon" definition in this document, which is
- * that a string can be considered "full entropy" if you can bound the min
- * entropy in each bit of output to at least 1-epsilon, where epsilon is
- * required to be <= 2^(-32).
+ * Set a callback to run on health failure in FIPS mode.
+ * This function will take an action determined by the caller.
  */
-#define ENTROPY_SAFETY_FACTOR		64
+typedef void (*jent_fips_failure_cb)(struct rand_data *ec,
+				     unsigned int health_failure);
+JENT_PRIVATE_STATIC
+int jent_set_fips_failure_callback(jent_fips_failure_cb cb);
+
+/* return version number of core library */
+JENT_PRIVATE_STATIC
+unsigned int jent_version(void);
+
 
 /**
  * Function pointer data structure to register an external thread handler
@@ -247,244 +207,12 @@ struct jent_notime_thread {
 	void (*jent_notime_stop)(void *ctx);
 };
 
-enum jent_startup_state {
-	jent_startup_completed,
-	jent_startup_sha3,
-	jent_startup_memory
-};
-
-/* The entropy pool */
-struct rand_data
-{
-	/* all data values that are vital to maintain the security
-	 * of the RNG are marked as SENSITIVE. A user must not
-	 * access that information while the RNG executes its loops to
-	 * calculate the next random value. */
-	void *hash_state;		/* SENSITIVE hash state entropy pool */
-	uint64_t prev_time;		/* SENSITIVE Previous time stamp */
-#define DATA_SIZE_BITS (JENT_SHA3_256_SIZE_DIGEST_BITS)
-
-#ifndef JENT_HEALTH_LAG_PREDICTOR
-	uint64_t last_delta;		/* SENSITIVE stuck test */
-	uint64_t last_delta2;		/* SENSITIVE stuck test */
-#endif /* JENT_HEALTH_LAG_PREDICTOR */
-
-	unsigned int flags;		/* Flags used to initialize */
-	unsigned int osr;		/* Oversampling rate */
-
-	/* Initialization state supporting AIS 20/31 NTG.1 */
-	enum jent_startup_state startup_state;
-
-#ifdef JENT_RANDOM_MEMACCESS
-  /* The step size should be larger than the cacheline size. */
-# ifndef JENT_MEMORY_BITS
-#  define JENT_MEMORY_BITS 17
-# endif
-# ifndef JENT_MEMORY_SIZE
-#  define JENT_MEMORY_SIZE (UINT32_C(1)<<JENT_MEMORY_BITS)
-# endif
-#else /* JENT_RANDOM_MEMACCESS */
-# ifndef JENT_MEMORY_BLOCKS
-#  define JENT_MEMORY_BLOCKS 512
-# endif
-# ifndef JENT_MEMORY_BLOCKSIZE
-#  define JENT_MEMORY_BLOCKSIZE 128
-# endif
-# ifndef JENT_MEMORY_SIZE
-#  define JENT_MEMORY_SIZE (JENT_MEMORY_BLOCKS*JENT_MEMORY_BLOCKSIZE)
-# endif
-#endif /* JENT_RANDOM_MEMACCESS */
-
-#define JENT_MEMORY_ACCESSLOOPS 128
-	unsigned char *mem;		/* Memory access location with size of
-					 * JENT_MEMORY_SIZE or memsize */
-#ifdef JENT_RANDOM_MEMACCESS
-	uint32_t memmask;		/* Memory mask (size of memory - 1) */
-#else
-	unsigned int memlocation; 	/* Pointer to byte in *mem */
-	unsigned int memblocks;		/* Number of memory blocks in *mem */
-	unsigned int memblocksize; 	/* Size of one memory block in bytes */
-#endif
-	unsigned int memaccessloops;	/* Number of memory accesses per random
-					 * bit generation */
-
-	/* Repetition Count Test */
-	int rct_count;			/* Number of stuck values */
-
-	/* Adaptive Proportion Test for a significance level of 2^-30 */
-	unsigned int apt_cutoff;	/* Intermittent health test failure */
-	unsigned int apt_cutoff_permanent; /* Permanent health test failure */
-#define JENT_APT_WINDOW_SIZE	512	/* Data window size */
-	unsigned int apt_observations;	/* Number of collected observations in
-					 * current window. */
-	unsigned int apt_count;		/* The number of times the reference
-					 * symbol been encountered in the
-					 * window. */
-	uint64_t apt_base;		/* APT base reference */
-	unsigned int health_failure;	/* Permanent health failure */
-
-	unsigned int apt_base_set:1;	/* APT base reference set? */
-	unsigned int fips_enabled:1;
-	unsigned int enable_notime:1;	/* Use internal high-res timer */
-	unsigned int max_mem_set:1;	/* Maximum memory configured by user */
-
-#ifdef JENT_CONF_ENABLE_INTERNAL_TIMER
-	volatile uint8_t notime_interrupt;	/* indicator to interrupt ctr */
-	volatile uint64_t notime_timer;		/* high-res timer mock-up */
-	uint64_t notime_prev_timer;		/* previous timer value */
-	void *notime_thread_ctx;		/* register thread data */
-#endif /* JENT_CONF_ENABLE_INTERNAL_TIMER */
-
-	uint64_t jent_common_timer_gcd;	/* Common divisor for all time deltas */
-
-#ifdef JENT_HEALTH_LAG_PREDICTOR
-	/* Lag predictor test to look for re-occurring patterns. */
-
-	/* The lag global cutoff selected based on the selection of osr. */
-	unsigned int lag_global_cutoff;
-
-	/* The lag local cutoff selected based on the selection of osr. */
-	unsigned int lag_local_cutoff;
-
-	/*
-	 * The number of times the lag predictor was correct. Compared to the
-	 * global cutoff.
-	 */
-	unsigned int lag_prediction_success_count;
-
-	/*
-	 * The size of the current run of successes. Compared to the local
-	 * cutoff.
-	 */
-	unsigned int lag_prediction_success_run;
-
-	/*
-	 * The total number of collected observations since the health test was
-	 * last reset.
-	 */
-	unsigned int lag_best_predictor;
-
-	/*
-	 * The total number of collected observations since the health test was
-	 * last reset.
-	 */
-	unsigned int lag_observations;
-
-	/*
-	 * This is the size of the window used by the predictor. The predictor
-	 * is reset between windows.
-	 */
-#define JENT_LAG_WINDOW_SIZE (1U<<17)
-
-	/*
-	 * The amount of history to base predictions on. This must be a power
-	 * of 2. Must be 4 or greater.
-	 */
-#define JENT_LAG_HISTORY_SIZE 8
-#define JENT_LAG_MASK (JENT_LAG_HISTORY_SIZE - 1)
-
-	/* The delta history for the lag predictor. */
-	uint64_t lag_delta_history[JENT_LAG_HISTORY_SIZE];
-
-	/* The scoreboard that tracks how successful each predictor lag is. */
-	unsigned int lag_scoreboard[JENT_LAG_HISTORY_SIZE];
-#endif /* JENT_HEALTH_LAG_PREDICTOR */
-};
-
-/* Flags that can be used to initialize the RNG */
-#define JENT_DISABLE_STIR (1<<0) 	/* UNUSED */
-#define JENT_DISABLE_UNBIAS (1<<1) 	/* UNUSED */
-#define JENT_DISABLE_MEMORY_ACCESS (1<<2) /* Disable memory access for more
-					     entropy, saves MEMORY_SIZE RAM for
-					     entropy collector */
-#define JENT_FORCE_INTERNAL_TIMER (1<<3)  /* Force the use of the internal
-					     timer */
-#define JENT_DISABLE_INTERNAL_TIMER (1<<4)  /* Disable the potential use of
-					       the internal timer. */
-#define JENT_FORCE_FIPS (1<<5)		  /* Force FIPS compliant mode
-					     including full SP800-90B
-					     compliance. */
-#define JENT_NTG1 (1<<6) /* AIS 20/31 NTG.1 compliance */
-
-/* Flags field limiting the amount of memory to be used for memory access */
-#define JENT_FLAGS_TO_MEMSIZE_SHIFT	28
-#define JENT_FLAGS_TO_MAX_MEMSIZE(val)	(val >> JENT_FLAGS_TO_MEMSIZE_SHIFT)
-#define JENT_MAX_MEMSIZE_TO_FLAGS(val)	(val << JENT_FLAGS_TO_MEMSIZE_SHIFT)
-#define JENT_MAX_MEMSIZE_32kB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 1))
-#define JENT_MAX_MEMSIZE_64kB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 2))
-#define JENT_MAX_MEMSIZE_128kB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 3))
-#define JENT_MAX_MEMSIZE_256kB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 4))
-#define JENT_MAX_MEMSIZE_512kB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 5))
-#define JENT_MAX_MEMSIZE_1MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 6))
-#define JENT_MAX_MEMSIZE_2MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 7))
-#define JENT_MAX_MEMSIZE_4MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 8))
-#define JENT_MAX_MEMSIZE_8MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C( 9))
-#define JENT_MAX_MEMSIZE_16MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C(10))
-#define JENT_MAX_MEMSIZE_32MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C(11))
-#define JENT_MAX_MEMSIZE_64MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C(12))
-#define JENT_MAX_MEMSIZE_128MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C(13))
-#define JENT_MAX_MEMSIZE_256MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C(14))
-#define JENT_MAX_MEMSIZE_512MB		JENT_MAX_MEMSIZE_TO_FLAGS(UINT32_C(15))
-#define JENT_MAX_MEMSIZE_MAX		JENT_MAX_MEMSIZE_512MB
-#define JENT_MAX_MEMSIZE_MASK		JENT_MAX_MEMSIZE_MAX
-/* We start at 32kB -> offset is log2(32768) */
-#define JENT_MAX_MEMSIZE_OFFSET		14
-
-#define JENT_MIN_OSR	3
-
-#ifdef JENT_PRIVATE_COMPILE
-# define JENT_PRIVATE_STATIC static
-#else /* JENT_PRIVATE_COMPILE */
-#if defined(_MSC_VER)
-#define JENT_PRIVATE_STATIC __declspec(dllexport)
-#else
-#define JENT_PRIVATE_STATIC __attribute__((visibility("default")))
-#endif
-#endif
-
-/* Number of low bits of the time value that we want to consider */
-/* get raw entropy */
-JENT_PRIVATE_STATIC
-ssize_t jent_read_entropy(struct rand_data *ec, char *data, size_t len);
-JENT_PRIVATE_STATIC
-ssize_t jent_read_entropy_safe(struct rand_data **ec, char *data, size_t len);
-/* initialize an instance of the entropy collector */
-JENT_PRIVATE_STATIC
-struct rand_data *jent_entropy_collector_alloc(unsigned int osr,
-	       				       unsigned int flags);
-/* clearing of entropy collector */
-JENT_PRIVATE_STATIC
-void jent_entropy_collector_free(struct rand_data *entropy_collector);
-
-/* initialization of entropy collector */
-JENT_PRIVATE_STATIC
-int jent_entropy_init(void);
-JENT_PRIVATE_STATIC
-int jent_entropy_init_ex(unsigned int osr, unsigned int flags);
-
-/*
- * Set a callback to run on health failure in FIPS mode.
- * This function will take an action determined by the caller.
- */
-typedef void (*jent_fips_failure_cb)(struct rand_data *ec,
-				     unsigned int health_failure);
-JENT_PRIVATE_STATIC
-int jent_set_fips_failure_callback(jent_fips_failure_cb cb);
-
-/* return version number of core library */
-JENT_PRIVATE_STATIC
-unsigned int jent_version(void);
-
 /* Set a different thread handling logic for the notimer support */
 JENT_PRIVATE_STATIC
 int jent_entropy_switch_notime_impl(struct jent_notime_thread *new_thread);
-
 /* -- END of Main interface functions -- */
 
 /* -- BEGIN timer-less threading support functions to prevent code dupes -- */
-
-#ifdef JENT_CONF_ENABLE_INTERNAL_TIMER
-
 struct jent_notime_ctx {
 	pthread_attr_t notime_pthread_attr;	/* pthreads library */
 	pthread_t notime_thread_id;		/* pthreads thread ID */
@@ -495,14 +223,6 @@ int jent_notime_init(void **ctx);
 
 JENT_PRIVATE_STATIC
 void jent_notime_fini(void *ctx);
-
-#else
-
-static inline int jent_notime_init(void **ctx) { (void)ctx; return 0; }
-static inline void jent_notime_fini(void *ctx) { (void)ctx; }
-
-#endif /* JENT_CONF_ENABLE_INTERNAL_TIMER */
-
 /* -- END timer-less threading support functions to prevent code dupes -- */
 
 /* -- BEGIN error codes for init function -- */