+++ src/jitterentropy-internal.h
@@ -50,8 +50,6 @@ extern "C" {
 
 #define ARRAY_SIZE(x) (sizeof(x) / sizeof((x)[0]))
 
-/* -- BEGIN Main interface functions -- */
-
 #ifndef JENT_STUCK_INIT_THRES
 /*
  * Per default, not more than 90% of all measurements during initialization
@@ -61,6 +59,279 @@ extern "C" {
  */
 #define JENT_STUCK_INIT_THRES(x) ((x*9) / 10)
 #endif
+
+/***************************************************************************
+ * Jitter RNG Configuration Section
+ *
+ * You may alter the following options
+ ***************************************************************************/
+
+/*
+ * Enable timer-less timer support with JENT_CONF_ENABLE_INTERNAL_TIMER
+ *
+ * In case the hardware is identified to not provide a high-resolution time
+ * stamp, this option enables a built-in high-resolution time stamp mechanism.
+ *
+ * The timer-less noise source is based on threads. This noise source requires
+ * the linking with the POSIX threads library. I.e. the executing environment
+ * must offer POSIX threads. If this option is disabled, no linking
+ * with the POSIX threads library is needed.
+ */
+
+/*
+ * Shall the LAG predictor health test be enabled?
+ */
+#define JENT_HEALTH_LAG_PREDICTOR
+
+/*
+ * Shall the jent_memaccess use a (statistically) random selection for the
+ * memory to update?
+ */
+#define JENT_RANDOM_MEMACCESS
+
+/*
+ * Mask specifying the number of bits of the raw entropy data of the time delta
+ * value used for the APT.
+ *
+ * This value implies that for the APT, only the bits specified by
+ * JENT_APT_MASK are taken. This was suggested in a draft IG D.K resolution 22
+ * provided by NIST, but further analysis
+ * (https://www.untruth.org/~josh/sp80090b/CMUF%20EWG%20Draft%20IG%20D.K%20Comments%20D10.pdf)
+ * suggests that this truncation / translation generally results in a health
+ * test with both a higher false positive rate (because multiple raw symbols
+ * map to the same symbol within the health test) and a lower statistical power
+ * when the APT cutoff is selected based on the apparent truncated entropy
+ * (i.e., truncation generally makes the test worse). NIST has since withdrawn
+ * this draft and stated that they will not propose truncation prior to
+ * health testing.
+ * Because the general tendency of such truncation to make the health test
+ * worse the default value is set such that no data is masked out and this
+ * should only be changed if a hardware-specific analysis suggests that some
+ * other mask setting is beneficial.
+ * The mask is applied to a time stamp where the GCD is already divided out, and thus no
+ * "non-moving" low-order bits are present.
+ */
+#define JENT_APT_MASK		(UINT64_C(0xffffffffffffffff))
+
+/* This parameter establishes the multiplicative factor that the desired
+ * memory region size should be larger than the observed cache size; the
+ * multiplicative factor is 2^JENT_CACHE_SHIFT_BITS.
+ * Set this to 0 if the desired memory region should be at least as large as
+ * the cache. If one wants most of the memory updates to result in a memory
+ * update, then this value should be at least 1.
+ * If the memory updates should dominantly result in a memory update, then
+ * the value should be set to at least 3.
+ * The actual size of the memory region is never larger than requested by
+ * the passed in JENT_MAX_MEMSIZE_* flag (if provided) or JENT_MEMORY_SIZE
+ * (if no JENT_MAX_MEMSIZE_* flag is provided).
+ */
+#ifndef JENT_CACHE_SHIFT_BITS
+#define JENT_CACHE_SHIFT_BITS 0
+#endif
+
+/*
+ * Meory access loop count: This value defines the default memory access loop
+ * count. The memory access loop is one of the hearts of the Jitter RNG. The
+ * number of loop counts has a direct impact on the entropy rate.
+ *
+ * It is permissible to configure this value differently at compile time if the
+ * observed entropy rate is too small.
+ *
+ * NOTE: When you modify this value, you are directly altering the behavior of
+ * the noise source. Make sure you fully understand what you do. If you want to
+ * individually measure the memory access loop entropy rate, use the
+ * jitterentropy-hashtime tool with the command line option of --memaccess.
+ */
+#ifndef JENT_MEM_ACC_LOOP_DEFAULT
+#define JENT_MEM_ACC_LOOP_DEFAULT 1
+#endif
+
+/*
+ * Hash loop count: This value defines the default hash loop count. The hash
+ * loop is one of the hearts of the Jitter RNG. The number of loop counts has a
+ * direct impact on the entropy rate.
+ *
+ * It is permissible to configure this value differently at compile time if the
+ * observed entropy rate is too small.
+ *
+ * NOTE: When you modify this value, you are directly altering the behavior of
+ * the noise source. Make sure you fully understand what you do. If you want to
+ * individually measure the memory access loop entropy rate, use the
+ * jitterentropy-hashtime tool with the command line option of --hashloop.
+ */
+#ifndef JENT_HASH_LOOP_DEFAULT
+#define JENT_HASH_LOOP_DEFAULT 1
+#endif
+
+/***************************************************************************
+ * Jitter RNG State Definition Section
+ ***************************************************************************/
+
+#define JENT_SHA3_256_SIZE_DIGEST_BITS	256
+#define JENT_SHA3_256_SIZE_DIGEST	(JENT_SHA3_256_SIZE_DIGEST_BITS >> 3)
+#define JENT_SHA3_512_SIZE_DIGEST_BITS	512
+#define JENT_SHA3_512_SIZE_DIGEST	(JENT_SHA3_512_SIZE_DIGEST_BITS >> 3)
+
+/*
+ * The output 256 bits can receive more than 256 bits of min entropy,
+ * of course, but the 256-bit output of SHA3-512_256(M) / XDRBG-256(M) can only
+ * asymptotically approach 256 bits of min entropy, not attain that bound.
+ * Random maps will tend to have output collisions, which reduces the creditable
+ * output entropy (that is what SP 800-90B Section 3.1.5.1.2 attempts to bound).
+ *
+ * The value "64" is justified in Appendix A.4 of the current 90C draft,
+ * and aligns with NIST's in "epsilon" definition in this document, which is
+ * that a string can be considered "full entropy" if you can bound the min
+ * entropy in each bit of output to at least 1-epsilon, where epsilon is
+ * required to be <= 2^(-32).
+ */
+#define ENTROPY_SAFETY_FACTOR		64
+
+enum jent_startup_state {
+	jent_startup_completed,
+	jent_startup_sha3,
+	jent_startup_memory
+};
+
+/* The entropy pool */
+struct rand_data
+{
+	/* all data values that are vital to maintain the security
+	 * of the RNG are marked as SENSITIVE. A user must not
+	 * access that information while the RNG executes its loops to
+	 * calculate the next random value. */
+	void *hash_state;		/* SENSITIVE hash state entropy pool */
+	uint64_t prev_time;		/* SENSITIVE Previous time stamp */
+#define DATA_SIZE_BITS (JENT_SHA3_256_SIZE_DIGEST_BITS)
+
+#ifndef JENT_HEALTH_LAG_PREDICTOR
+	uint64_t last_delta;		/* SENSITIVE stuck test */
+	uint64_t last_delta2;		/* SENSITIVE stuck test */
+#endif /* JENT_HEALTH_LAG_PREDICTOR */
+
+	unsigned int flags;		/* Flags used to initialize */
+	unsigned int osr;		/* Oversampling rate */
+
+	/* Initialization state supporting AIS 20/31 NTG.1 */
+	enum jent_startup_state startup_state;
+
+#ifdef JENT_RANDOM_MEMACCESS
+  /* The step size should be larger than the cacheline size. */
+# ifndef JENT_MEMORY_BITS
+#  define JENT_MEMORY_BITS 17
+# endif
+# ifndef JENT_MEMORY_SIZE
+#  define JENT_MEMORY_SIZE (UINT32_C(1)<<JENT_MEMORY_BITS)
+# endif
+#else /* JENT_RANDOM_MEMACCESS */
+# ifndef JENT_MEMORY_BLOCKS
+#  define JENT_MEMORY_BLOCKS 512
+# endif
+# ifndef JENT_MEMORY_BLOCKSIZE
+#  define JENT_MEMORY_BLOCKSIZE 128
+# endif
+# ifndef JENT_MEMORY_SIZE
+#  define JENT_MEMORY_SIZE (JENT_MEMORY_BLOCKS*JENT_MEMORY_BLOCKSIZE)
+# endif
+#endif /* JENT_RANDOM_MEMACCESS */
+
+#define JENT_MEMORY_ACCESSLOOPS 128
+	unsigned char *mem;		/* Memory access location with size of
+					 * JENT_MEMORY_SIZE or memsize */
+#ifdef JENT_RANDOM_MEMACCESS
+	uint32_t memmask;		/* Memory mask (size of memory - 1) */
+#else
+	unsigned int memlocation; 	/* Pointer to byte in *mem */
+	unsigned int memblocks;		/* Number of memory blocks in *mem */
+	unsigned int memblocksize; 	/* Size of one memory block in bytes */
+#endif
+	unsigned int memaccessloops;	/* Number of memory accesses per random
+					 * bit generation */
+
+	/* Repetition Count Test */
+	int rct_count;			/* Number of stuck values */
+
+	/* Adaptive Proportion Test for a significance level of 2^-30 */
+	unsigned int apt_cutoff;	/* Intermittent health test failure */
+	unsigned int apt_cutoff_permanent; /* Permanent health test failure */
+#define JENT_APT_WINDOW_SIZE	512	/* Data window size */
+	unsigned int apt_observations;	/* Number of collected observations in
+					 * current window. */
+	unsigned int apt_count;		/* The number of times the reference
+					 * symbol been encountered in the
+					 * window. */
+	uint64_t apt_base;		/* APT base reference */
+	unsigned int health_failure;	/* Permanent health failure */
+
+	unsigned int apt_base_set:1;	/* APT base reference set? */
+	unsigned int fips_enabled:1;
+	unsigned int enable_notime:1;	/* Use internal high-res timer */
+	unsigned int max_mem_set:1;	/* Maximum memory configured by user */
+
+#ifdef JENT_CONF_ENABLE_INTERNAL_TIMER
+	volatile uint8_t notime_interrupt;	/* indicator to interrupt ctr */
+	volatile uint64_t notime_timer;		/* high-res timer mock-up */
+	uint64_t notime_prev_timer;		/* previous timer value */
+	void *notime_thread_ctx;		/* register thread data */
+#endif /* JENT_CONF_ENABLE_INTERNAL_TIMER */
+
+	uint64_t jent_common_timer_gcd;	/* Common divisor for all time deltas */
+
+#ifdef JENT_HEALTH_LAG_PREDICTOR
+	/* Lag predictor test to look for re-occurring patterns. */
+
+	/* The lag global cutoff selected based on the selection of osr. */
+	unsigned int lag_global_cutoff;
+
+	/* The lag local cutoff selected based on the selection of osr. */
+	unsigned int lag_local_cutoff;
+
+	/*
+	 * The number of times the lag predictor was correct. Compared to the
+	 * global cutoff.
+	 */
+	unsigned int lag_prediction_success_count;
+
+	/*
+	 * The size of the current run of successes. Compared to the local
+	 * cutoff.
+	 */
+	unsigned int lag_prediction_success_run;
+
+	/*
+	 * The total number of collected observations since the health test was
+	 * last reset.
+	 */
+	unsigned int lag_best_predictor;
+
+	/*
+	 * The total number of collected observations since the health test was
+	 * last reset.
+	 */
+	unsigned int lag_observations;
+
+	/*
+	 * This is the size of the window used by the predictor. The predictor
+	 * is reset between windows.
+	 */
+#define JENT_LAG_WINDOW_SIZE (1U<<17)
+
+	/*
+	 * The amount of history to base predictions on. This must be a power
+	 * of 2. Must be 4 or greater.
+	 */
+#define JENT_LAG_HISTORY_SIZE 8
+#define JENT_LAG_MASK (JENT_LAG_HISTORY_SIZE - 1)
+
+	/* The delta history for the lag predictor. */
+	uint64_t lag_delta_history[JENT_LAG_HISTORY_SIZE];
+
+	/* The scoreboard that tracks how successful each predictor lag is. */
+	unsigned int lag_scoreboard[JENT_LAG_HISTORY_SIZE];
+#endif /* JENT_HEALTH_LAG_PREDICTOR */
+};
+
+#define JENT_MIN_OSR	3
 
 #ifdef __cplusplus
 }